feat: 메소드 분리 및 자바독 작성
